SI-1264 make CompileSocket send the right things to stderr/stdout
This properly splits the output of fsc (from `scala -e`, scala worksheet,
etc.) between stderr (for warnings and errors) and stdout (for the
rest). With this change, fsc behaves the same as `scalac`.

This makes the fsc server prefix messages with the destination
([err]/[out]), and the client socket strip them and send the rest to the
client stderr/stdout.