Eliminate unreachable code in GenBCode
We rely on dead code elimination provided by the ASM framework, as
described in the ASM User Guide (http://asm.ow2.org/index.html),
Section 8.2.1. It runs a data flow analysis, which only computes
information for reachable instructions. Instructions for which no
data is available after the analyis are unreachable.

There's one issue with the ASM framework (or the way we use it): Data
flow analysis requires the maxlocals and maxstack of each method to be
computed. The ASM framework calculates these maxes only when
writing the classfiles, not during code generation. In order to run
DCE, we therefore run a MethodWriter beforehand on every method. This
assings the MethodNode's maxStack/maxLocals, but it does more work
(writes the instructions to a byte array).

This is also what Miguel uses on his branch. The change is basically
the same as https://github.com/lrytz/scala/commit/bfadf92c20.

We could probably make this faster (and allocate less memory) by
hacking the ASM framework: create a subclass of MethodWriter with a
/dev/null byteVector. Another option would be to create a separate
visitor for computing those values, duplicating the functionality from
the MethodWriter.

For now, I added some timers to be able to measure the time DCE takes.
Here's compiling the library with -Ystatistics:jvm

    time in backend               : 1 spans, 6597ms
    bcode initialization          : 1 spans, 8ms (0.1%)
    code generation               : 1 spans, 4580ms (69.4%)
    dead code elimination         : 3771 spans, 742ms (11.2%)
    classfile writing             : 1 spans, 879ms (13.3%)