SI-8828 fix regression in Xlint visibility warning for sealed classes
5dfcf5e reverted a change to `Symbol#isEffectivelyFinal` (made in
adeffda) that broke overriding checks, and moved the new enhanced
version to a new method.

However, the test for inaccessible type access still uses the old one,
so it lost the ability to see that the owner of some method is either
final or sealed and not overridden.

This just makes it use the new `isEffectivelyFinalOrNotOverriden`.