SI-8410 Summarize if warnings and not disabled
Simplify the conditional and emit the summary if
there were buffered warnings and the user didn't
explicitly disable this setting.

The working assumption is that using fatal-warnings
to turn on summaries was either an outdated heuristic
or just due to a faulty merge. Or wait, was it for
-feature, which is not enabled when warnings are
issued against it?