Particular formatting improvement in the spec
Fixes the titularity of the "Overloading resolution" header
under 6.26.1 "Value conversions" to match "Type instantiation."

Hopefully this makes the link to the actual 6.26.3 overloading
resolution section work without further renaming. (Edit: changed
the minor heading to say "Static overload resolution", which is
sufficiently redundant.)

Possibly that is the oftenest-cited section of the entire document.

At least on Stack Overflow, if not in Proceedings of the ACM.