SI-8927 Update OSGi stuff to get rid of bndlib warning
The way the desired OSGi frameworks are chosen has changed between
Pax Exam versions:

On earlier versions, specifying it in code was fine, but PAX Exam 4.x
runs the tests on the first OSGi framework it finds on classpath.

An exclusion for the transitive dependency org.osgi.core was added,
because it seems that artifact has broken dependecies (which would
have brought us back to square one).
Good thing is that it isn't needed here, because the OSGi framework JARs
contain all necessary classes.