SI-8922 REPL load -v
Verbose mode causes the familiar prompt and
line echo so you can see what you just loaded.

The quit message is pushed up a level in the
process loop.

This has the huge payoff that if you start the
repl and immediately hit ctl-D, you don't have to
wait for the compiler to init (yawn) before you
get a shell prompt back.