Proper cleanup after finishing global-showdef test
The temporary foo directory with classes created by global-showdef.scala was not deleted after finishing a test. When I was running tests from Scala project root, this directory was created there and was visible when running git status.

I just forced the creation of a foo directory inside global-showdef-run.obj - so it's properly deleted after finishing the test.