SI-8898 javap -fun under new style lambdas
To support both -Ydelambdafy strategies, look for both inline
(anonfun) and method (lambda) closure classes.

For method (lambda) style, use the anonfun method that is
invoked by the accessor.

Also, the output of javap must be captured eagerly for
filtering for the current target method.

If the user asks for a module, e.g., `Foo$`, don't yield
results for companion class, but for `Foo`, do yield
companion module results.  Just because.